# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/scim/v2/Users": {
      "get": {
        "operationId": "ScimController_listUsers",
        "parameters": [
          {
            "name": "startIndex",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "post": {
        "operationId": "ScimController_createUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      }
    },
    "/scim/v2/Users/{id}": {
      "get": {
        "operationId": "ScimController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "put": {
        "operationId": "ScimController_replaceUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "patch": {
        "operationId": "ScimController_patchUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "delete": {
        "operationId": "ScimController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      }
    },
    "/scim/v2/Groups": {
      "get": {
        "operationId": "ScimController_listGroups",
        "parameters": [
          {
            "name": "startIndex",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "post": {
        "operationId": "ScimController_createGroup",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      }
    },
    "/scim/v2/Groups/{id}": {
      "get": {
        "operationId": "ScimController_getGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "patch": {
        "operationId": "ScimController_patchGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      },
      "delete": {
        "operationId": "ScimController_deleteGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Scim"]
      }
    },
...
  },
  "info": {
    "title": "HandStack API",
    "description": "Versioned public and administrative HandStack API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}